feat: Update stellar custom.yaml configuration
The commit updates the `custom.yaml` configuration file for the Stellar module. It modifies the `epochs` value from 5 to 20, the `weight_decay` value from 0.001 to 0.0001, the `hid_dim` value from 50 to 160, and sets `batch_norm` to false. Additionally, it changes the `n_graph_layers` value to 1 and updates the `graph_conv_impl` to use the `FiLMConv` implementation.